requirements.txt
tensorflow>=2.16.0
numpy>=1.23.0
opencv-python>=4.5.0
pillow>=9.0.0
customtkinter>=5.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.5.0
h5py>=3.0.0
protobuf>=4.0.0,<5.0.0